feat: Add support for dark mode (#100)
Added support for dark mode, thanks to @CodingByJerez! 🙌
The dialog will now automatically switch to a dark theme when the device is on dark-mode. 

BREAKING CHANGE: We are now using the native colors instead of the hardcoded ones — which means that if your react-native app is using a specific accent/primary color it will be now shown correctly in the dialog.