Permit updating the value of code inputs
Prior to this commit, there is no way to programmatically update the
value of a CodeInput.

After this commit, sending a value prop to the CodeInput component
allows us to control the value displayed in the CodeInput, by changing
the content of the variable that's set as the CodeInput's value prop.